refactor(calendar): extract time calculatation
