+++ src/composables/useCalendar.ts
@@ -4,27 +4,13 @@ import { useGoogleCalendar } from './useGoogleCalendar';
 import { useTimeUtilities } from './useTimeUtilities';
 
 const selectedEvents = ref<EventApi[]>([]);
-
-const {
-  workDayIntervals,
-  workTimeIntervals,
-  createEventInterval,
-  calcIntervalsDifference,
-  intervalsToHours,
-} = useTimeUtilities();
-
-const eventsIntervals = computed(() => selectedEvents.value.map((event) => createEventInterval(event)));
-const devTimeIntervals = computed(() => calcIntervalsDifference(workTimeIntervals.value, eventsIntervals.value));
-
-const workHours = computed(() => intervalsToHours(workTimeIntervals.value));
-const devHours = computed(() => intervalsToHours(devTimeIntervals.value));
-
 const toggleSelectedEvent = (event :EventApi) => {
   selectedEvents.value = selectedEvents.value.some((x) => x.id === event.id)
     ? selectedEvents.value.filter((x) => x.id !== event.id)
     : [...selectedEvents.value, event];
 };
 
+const { workDayIntervals } = useTimeUtilities();
 const { events: _events } = useGoogleCalendar();
 const events = computed(() => [
   ..._events.value.map<EventInput>((event) => ({
@@ -57,6 +43,4 @@ export const useCalendar = () => ({
   events,
   selectedEvents,
   toggleSelectedEvent,
-  workHours,
-  devHours,
 });
